requirements.txt
numpy>=1.24
Flask>=3.0
matplotlib>=3.8
Pillow>=10.0